Refactored pmt_wrap.
Pmt_wrap used to sit on the top of the pmt hierarchy.  It now sits just above
pmt_base.  Several of the functions are templatized and specialized in the
various header files.  This means that we can add a new pmt type without
pmt_wrap having to know about it.

Signed-off-by: John Sallay <jasallay@gmail.com>