meson: do not require cmake when finding dependencies
fmt and spdlog provide pkg-config files, and these work and compile
perfectly fine. Meson can try finding with both pkg-config and cmake,
if it is allowed to, which means that dependency lookup now works on
systems where cmake is not installed.

Even when using cmake for dependency lookup, the modules kwarg is not
needed. Both dependencies export a single cmake target, so Meson can
figure out the correct one to use without trouble.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: John Sallay <jasallay@gmail.com>