pmt: use libdir setting for python and pkgconfig path
At the moment the libdir setting only influences the location of libpmtf.so.
python(xx) and pkgconfig always go to lib independently of the libdir setting.

This pr evaluates the libdir setting and also uses it for the location
of the python(xx) and pkgconfig directories.

This is the same behaviour as in gnuradio 3.x

Python, cmake, pkgconfig all go under lib(64). There is no mixture between lib and lib64.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>