Rewrite of PMTs to use variants.
Some aspects of flatbuffers are really slow (like construction) and we
weren't really using flatbuffers in the way that they were intended
anyway.

This is a complete rewrite of the PMTs using std::variants.  The code is
much simpler and the performance is better (and in some cases an order
of magnitude better).

The code now requires c++20 as well.

Signed-off-by: John Sallay <jasallay@gmail.com>