Fmt support (#80)
ostream operations are being supplanted by fmtlib (std::format). Adds support for the fmt functions with unit tests.
Replaces ostream functions with a call to fmt::format.

Signed-off-by: John Sallay <jasallay@gmail.com>

* Add missing includes and install base64 to right place

Signed-off-by: Josh Morman <jmorman@gnuradio.org>
Signed-off-by: John Sallay <jasallay@gmail.com>

* Remove debug print in buffer_to_pmt

Signed-off-by: Josh Morman <jmorman@gnuradio.org>
Signed-off-by: John Sallay <jasallay@gmail.com>

* Support for fmt is tested.

Replace ostream functions with a call to fmt::format.

Signed-off-by: John Sallay <jasallay@gmail.com>

* Move function def out of class to try to fix compile issue

Signed-off-by: John Sallay <jasallay@gmail.com>

* Fix gcc11 recursive format issue.

Due to an issue/spec ambiguity we can't have a recursive formatter for
pmts.  Found a way around this using nested lambda functions.  Not my
favorite but I believe that it is required.

Signed-off-by: John Sallay <jasallay@gmail.com>

---------

Signed-off-by: John Sallay <jasallay@gmail.com>
Signed-off-by: Josh Morman <jmorman@gnuradio.org>
Co-authored-by: Josh Morman <jmorman@gnuradio.org>